Fix get all restaurants + add get restaurant menu

+++ app/src/main/java/com/example/deliveryfoodapp/pages/validate_payment_page.kt
@@ -1,7 +1,10 @@
 package com.example.deliveryfoodapp.pages
 
+import android.annotation.SuppressLint
+import android.widget.Toast
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
@@ -12,12 +15,14 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableDoubleStateOf
 import androidx.compose.runtime.mutableIntStateOf
@@ -27,6 +32,7 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
@@ -35,6 +41,7 @@ import androidx.compose.ui.unit.sp
 import androidx.navigation.NavHostController
 import com.example.deliveryfoodapp.R
 import com.example.deliveryfoodapp.authenticatedUser
+import com.example.deliveryfoodapp.backend_services.restaurant_api.RestaurantEndpoints
 import com.example.deliveryfoodapp.currentRestaurant
 import com.example.deliveryfoodapp.models.UserCart
 import com.example.deliveryfoodapp.local_storage_services.repositories.UserCartRepository
@@ -45,9 +52,13 @@ import com.example.deliveryfoodapp.utils.Routes
 import com.example.deliveryfoodapp.widgets.CustomAppBar
 import com.example.deliveryfoodapp.widgets.PrincipalButton
 
+@SuppressLint("DefaultLocale")
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun ValidatePaymentPage(navController : NavHostController) {
+
+    val context = LocalContext.current
+    val isLoading = remember { mutableStateOf(false) }
 
     var userCart = authenticatedUser.getUserCartByRestaurantID(restaurantID = currentRestaurant.id)
 
@@ -297,7 +308,7 @@ fun ValidatePaymentPage(navController : NavHostController) {
                             )
 
                             Text(
-                                text = "$Items_total DA",
+                                text = "${String.format("%.1f",Items_total)} DA",
                                 style = TextStyle(
                                     fontSize = 13.sp,
                                     fontWeight = FontWeight.SemiBold
@@ -362,7 +373,7 @@ fun ValidatePaymentPage(navController : NavHostController) {
                             )
 
                             Text(
-                                text = "$Total_price DA",
+                                text = "${String.format("%.1f",Total_price)} DA",
                                 style = TextStyle(
                                     fontSize = 16.sp,
                                     fontWeight = FontWeight.Bold,
@@ -386,28 +397,38 @@ fun ValidatePaymentPage(navController : NavHostController) {
             }
         }
 
-        PrincipalButton(
-            text = "Place the order",
-            onClick = {
-
-                // TODO Create new order fel backend
-
-                // Delete that cart from the authenticatedUser
-                authenticatedUser.deleteCart(userCart)
-
-                // delete that cart from SqlLite (their orderItems will be deleted automatically)
-                val roomUserCart : RoomUserCart? = UserCartRepository.getUserCartById(userCart.id)
-                if (roomUserCart != null) {
-                    UserCartRepository.removeUserCart(roomUserCart)
-                }
-                // Empty userCart variable
-                userCart = UserCart.emptyUserCart()
-
-
-                navController.navigate(Routes.ORDER_PLACED_PAGE) {
-                    popUpTo(0) { inclusive = true } // Clear the entire back stack
-                }
+        if (isLoading.value) {
+            Box(
+                modifier = Modifier.fillMaxWidth()
+            ){
+                CircularProgressIndicator(modifier = Modifier.align(Alignment.Center))
             }
-        )
+        }else {
+            PrincipalButton(
+                text = "Place the order",
+                onClick = {
+
+                    // TODO Create new order fel backend
+                    isLoading.value = true
+                    // TODO : Call the backend function, and then when it complete with sucess, turn isLoading into false
+
+                    // Delete that cart from the authenticatedUser
+                    authenticatedUser.deleteCart(userCart)
+
+                    // delete that cart from SqlLite (their orderItems will be deleted automatically)
+                    val roomUserCart : RoomUserCart? = UserCartRepository.getUserCartById(userCart.id)
+                    if (roomUserCart != null) {
+                        UserCartRepository.removeUserCart(roomUserCart)
+                    }
+                    // Empty userCart variable
+                    userCart = UserCart.emptyUserCart()
+
+
+                    navController.navigate(Routes.ORDER_PLACED_PAGE) {
+                        popUpTo(0) { inclusive = true } // Clear the entire back stack
+                    }
+                }
+            )
+        }
     }
 }
